third_party/blink/web_tests/external/wpt/css/css-fonts/variations/font-style-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] Font-style (computed): 'oblique' is valid
  assert_equals: Font-style computed style: 'oblique' is valid expected "oblique" but got "italic"
[FAIL] Font-style (computed): 'oblique' followed by calc is valid even if it mixes units (with computation)
  assert_equals: Font-style computed style: 'oblique' followed by calc is valid even if it mixes units (with computation) expected "oblique" but got "italic"
Harness: the test ran to completion.
